Moved precompiler functionality from scripts folder into src folder.
